+++ src/label.h
@@ -25,8 +25,8 @@ namespace gdstk {
 enum struct Anchor { NW = 0, N = 1, NE = 2, W = 4, O = 5, E = 6, SW = 8, S = 9, SE = 10 };
 
 struct Label {
-    int16_t layer;
-    int16_t texttype;
+    uint32_t layer;
+    uint32_t texttype;
     char* text;  // 0-terminated
     Vec2 origin;
     Anchor anchor;
